Issue #2992817 by RoSk0, johndevman, phenaproxima, tim.plunkett, yanniboi, alexpott: Layout builder cron errors when no cleanup required
(cherry picked from commit 964372f57a6af16b3d04fef545450d48f75ede41)

+++ core/modules/layout_builder/src/InlineBlockUsage.php
@@ -87,8 +87,10 @@ class InlineBlockUsage {
    *   The block content entity IDs.
    */
   public function deleteUsage(array $block_content_ids) {
-    $query = $this->database->delete('inline_block_usage')->condition('block_content_id', $block_content_ids, 'IN');
-    $query->execute();
+    if (!empty($block_content_ids)) {
+      $query = $this->database->delete('inline_block_usage')->condition('block_content_id', $block_content_ids, 'IN');
+      $query->execute();
+    }
   }
 
   /**
